+++ curvlinops/fisher.py
@@ -195,8 +195,8 @@ class FisherMCLinearOperator(CurvatureLinearOperator):
         Returns:
             Matrix-multiplication result ``mat @ M`` in tensor list format.
         """
-        if self._generator is None or self._generator.device != self._device:
-            self._generator = Generator(device=self._device)
+        if self._generator is None or self._generator.device != self.device:
+            self._generator = Generator(device=self.device)
         self._generator.manual_seed(self._seed)
 
         return super()._matmat(M)