+++ docs/examples/basic_usage/example_matrix_vector_products.py
@@ -9,17 +9,14 @@ First, the imports.
 """
 
 import matplotlib.pyplot as plt
-import numpy
-import torch
-from torch import nn
+from torch import cat, cuda, device, eye, manual_seed, nn, rand
 
 from curvlinops import GGNLinearOperator, HessianLinearOperator
 from curvlinops.examples.functorch import functorch_ggn, functorch_hessian
-from curvlinops.examples.utils import report_nonclose
+from curvlinops.utils import allclose_report
 
 # make deterministic
-torch.manual_seed(0)
-numpy.random.seed(0)
+manual_seed(0)
 
 # %%
 # Setup
@@ -31,10 +28,10 @@ D_in = 7
 D_hidden = 5
 D_out = 3
 
-DEVICE = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+DEVICE = device("cuda" if cuda.is_available() else "cpu")
 
-X = torch.rand(N, D_in).to(DEVICE)
-y = torch.rand(N, D_out).to(DEVICE)
+X = rand(N, D_in, device=DEVICE)
+y = rand(N, D_out, device=DEVICE)
 
 model = nn.Sequential(
     nn.Linear(D_in, D_hidden),
@@ -55,16 +52,14 @@ loss_function = nn.MSELoss(reduction="mean").to(DEVICE)
 # Setting up a linear operator for the Hessian is straightforward.
 
 data = [(X, y)]
-H = HessianLinearOperator(model, loss_function, params, data).to_scipy()
+H = HessianLinearOperator(model, loss_function, params, data)
 
 # %%
 #
-# We can now multiply by the Hessian. This operation will be carried out in
-# PyTorch under the hood, but the operator is compatible with ``scipy``, so we
-# can just pass a ``numpy`` vector to the matrix-multiplication.
+# We can now multiply the Hessian onto a vector.
 
 D = H.shape[0]
-v = numpy.random.rand(D)
+v = rand(D, device=DEVICE)
 
 Hv = H @ v
 
@@ -73,7 +68,7 @@ Hv = H @ v
 # To verify the result, we compute the Hessian using ``functorch``, using a
 # utility function from ``curvlinops.examples``:
 
-H_mat = functorch_hessian(model, loss_function, params, data).detach().cpu().numpy()
+H_mat = functorch_hessian(model, loss_function, params, data).detach()
 
 # %%
 #
@@ -82,7 +77,7 @@ H_mat = functorch_hessian(model, loss_function, params, data).detach().cpu().num
 Hv_functorch = H_mat @ v
 
 print("Comparing Hessian-vector product with functorch's Hessian-vector product.")
-report_nonclose(Hv, Hv_functorch)
+assert allclose_report(Hv, Hv_functorch)
 
 
 # %%
@@ -92,14 +87,14 @@ report_nonclose(Hv, Hv_functorch)
 # We can also compute the Hessian's matrix representation with the linear
 # operator, simply by multiplying it onto the identity matrix. (Of course, this
 # only works if the Hessian is small enough.)
-H_mat_from_linop = H @ numpy.eye(D)
+H_mat_from_linop = H @ eye(D, device=DEVICE)
 
 # %%
 #
 # This should yield the same matrix as with :code:`functorch`.
 
 print("Comparing Hessian with functorch's Hessian.")
-report_nonclose(H_mat, H_mat_from_linop)
+assert allclose_report(H_mat, H_mat_from_linop)
 
 # %%
 #
@@ -111,19 +106,82 @@ plt.imshow(H_mat)
 plt.colorbar()
 
 # %%
+# Accepted vector/matrix formats
+# ------------------------------
+#
+# Curvature matrices are usually defined w.r.t. parameters of a neural net. In PyTorch,
+# these parameters are split into multiple tensors (e.g. per layer). It is often more
+# convenient to think and work with vectors/matrices defined in this list format, rather
+# than in the flattened-and-concatenated parameter space.
+#
+# So far, we have only used vectors/matrices in the flattened-and-concatenated format.
+# To account for the often more convenient list format, all linear operators in
+# ``curvlinops`` can also handle vectors/matrices specified in tensor list format.
+# In that format, a matrix is a list of tensors, each of which has the same shape as
+# its corresponding parameter plus an additional trailing dimension for the matrix's
+# column dimension.
+#
+# ``curvlinops`` preserves the format when performing matrix multiplies: If the input
+# lived in the flattened-and-concatenated parameter space, the result will be as well.
+# If the input lived in the tensor list parameter space, the result will be a tensor
+# list as well.
+#
+# Let's make this concrete. First, set up the same matrix in flattened and list format:
+
+num_columns = 3
+
+print(f"Total network parameters: {D}")
+print(f"Parameter shapes: {[p.shape for p in params]}")
+print(f"Number of columns: {num_columns}")
+
+# Matrix in tensor list format
+M_list = [rand(*p.shape, num_columns, device=DEVICE) for p in params]
+print(f"[Tensor list format] Matrix: {[m.shape for m in M_list]}")
+
+# Matrix in flattened format (what we have been using before)
+M_flat = cat([m.flatten(end_dim=-2) for m in M_list])
+print(f"[Flat format] Matrix: {M_flat.shape}")
+
+# %%
+#
+# Next, let's carry out the Hessian-matrix product and inspect the result's format:
+
+HM_list = H @ M_list
+print(f"[Tensor list format] Hessian-matrix product: {[hm.shape for hm in HM_list]}")
+
+HM_flat = H @ M_flat
+print(f"[Flat format] Hessian-matrix product: {HM_flat.shape}")
+
+# %%
+#
+# As expected, this produces the same result:
+
+HM_list_flattened = cat([hm.flatten(end_dim=-2) for hm in HM_list])
+
+print("Comparing Hessian-matrix products across formats.")
+assert allclose_report(HM_flat, HM_list_flattened)
+
+# %%
+#
+# **Note:** Like in the early part of the tutorial, the column dimension is not
+# necessary if we just want to multiply the Hessian onto a single vector.
+#
 # GGN-vector products
 # -------------------
 #
 # Setting up a linear operator for the Fisher/GGN is identical to the Hessian.
 
-GGN = GGNLinearOperator(model, loss_function, params, data).to_scipy()
+GGN = GGNLinearOperator(model, loss_function, params, data)
 
 # %%
+#
+# This is one of ``curvlinops``'s design features: All linear operators share the same
+# interface, making it easy to switch between curvature matrices.
 #
 # Let's compute a GGN-vector product.
 
 D = H.shape[0]
-v = numpy.random.rand(D)
+v = rand(D, device=DEVICE)
 
 GGNv = GGN @ v
 
@@ -134,12 +192,12 @@ GGNv = GGN @ v
 # network by its linearization. This is implemented in a utility function of
 # :code:`curvlinops.examples`:
 
-GGN_mat = functorch_ggn(model, loss_function, params, data).detach().cpu().numpy()
+GGN_mat = functorch_ggn(model, loss_function, params, data).detach()
 
 GGNv_functorch = GGN_mat @ v
 
 print("Comparing GGN-vector product with functorch's GGN-vector product.")
-report_nonclose(GGNv, GGNv_functorch)
+assert allclose_report(GGNv, GGNv_functorch)
 
 # %%
 # GGN-matrix products
@@ -148,14 +206,14 @@ report_nonclose(GGNv, GGNv_functorch)
 # We can also compute the GGN matrix representation with the linear operator,
 # simply by multiplying it onto the identity matrix. (Of course, this only
 # works if the GGN is small enough.)
-GGN_mat_from_linop = GGN @ numpy.eye(D)
+GGN_mat_from_linop = GGN @ eye(D, device=DEVICE)
 
 # %%
 #
 # This should yield the same matrix as with :code:`functorch`.
 
 print("Comparing GGN with functorch's GGN.")
-report_nonclose(GGN_mat, GGN_mat_from_linop)
+assert allclose_report(GGN_mat, GGN_mat_from_linop)
 
 # %%
 #