+++ test/diagonal/test_epperly2024xtrace.py
@@ -1,20 +1,20 @@
 """Test ``curvlinops.diagonal.epperly2024xtrace``."""
 
 from functools import partial
+from typing import Union
 
-from numpy import allclose, column_stack, diag, mean, ndarray
-from numpy.linalg import qr
-from numpy.random import rand, seed
 from pytest import mark
-from scipy.sparse.linalg import LinearOperator
+from torch import Tensor, column_stack, manual_seed, rand
+from torch.linalg import qr
 
 from curvlinops import xdiag
+from curvlinops._torch_base import PyTorchLinearOperator
 from curvlinops.sampling import random_vector
 from test.diagonal import NUM_MATVEC_IDS, NUM_MATVECS
 from test.utils import check_estimator_convergence
 
 
-def xdiag_naive(A: LinearOperator, num_matvecs: int) -> ndarray:
+def xdiag_naive(A: Union[PyTorchLinearOperator, Tensor], num_matvecs: int) -> Tensor:
     """Naive reference implementation of XDiag.
 
     See Section 2.4 in https://arxiv.org/pdf/2301.07825.
@@ -42,7 +42,9 @@ def xdiag_naive(A: LinearOperator, num_matvecs: int) -> ndarray:
         )
     num_vecs = num_matvecs // 2
 
-    W = column_stack([random_vector(dim, "rademacher") for _ in range(num_vecs)])
+    W = column_stack(
+        [random_vector(dim, "rademacher", A.device, A.dtype) for _ in range(num_vecs)]
+    )
     A_W = A @ W
 
     diagonals = []
@@ -53,7 +55,7 @@ def xdiag_naive(A: LinearOperator, num_matvecs: int) -> ndarray:
         not_i = [j for j in range(num_vecs) if j != i]
         Q_i, _ = qr(A_W[:, not_i])
         QT_i_A = Q_i.T @ A
-        diag_Q_i_QT_i_A = diag(Q_i @ QT_i_A)
+        diag_Q_i_QT_i_A = (Q_i @ QT_i_A).diag()
 
         # apply vanilla Hutchinson in the complement, using test vector i
         w_i = W[:, i]
@@ -61,7 +63,7 @@ def xdiag_naive(A: LinearOperator, num_matvecs: int) -> ndarray:
         diag_w_i = w_i * (A_w_i - Q_i @ (Q_i.T @ A_w_i)) / w_i**2
         diagonals.append(diag_Q_i_QT_i_A + diag_w_i)
 
-    return mean(diagonals, axis=0)
+    return sum(diagonals) / len(diagonals)
 
 
 @mark.parametrize("num_matvecs", NUM_MATVECS, ids=NUM_MATVEC_IDS)
@@ -71,11 +73,11 @@ def test_xdiag(num_matvecs: int):
     Args:
         num_matvecs: Number of matrix-vector multiplications used by one estimator.
     """
-    seed(0)
+    manual_seed(0)
     A = rand(30, 30)
 
     estimator = partial(xdiag, A=A, num_matvecs=num_matvecs)
-    check_estimator_convergence(estimator, num_matvecs, diag(A), target_rel_error=3e-2)
+    check_estimator_convergence(estimator, num_matvecs, A.diag(), target_rel_error=3e-2)
 
 
 @mark.parametrize("num_matvecs", NUM_MATVECS, ids=NUM_MATVEC_IDS)
@@ -87,13 +89,13 @@ def test_xdiag_matches_naive(num_matvecs: int, num_seeds: int = 5):
         num_seeds: Number of different seeds to test the estimators with.
             Default: ``5``.
     """
-    seed(0)
-    A = rand(30, 30)
+    manual_seed(0)
+    A = rand(30, 30).double()
 
     # check for different seeds
     for i in range(num_seeds):
-        seed(i)
+        manual_seed(i)
         efficient = xdiag(A, num_matvecs)
-        seed(i)
+        manual_seed(i)
         naive = xdiag_naive(A, num_matvecs)
-        assert allclose(efficient, naive)
+        assert efficient.allclose(naive)