+++ test/test_ekfac.py
@@ -81,7 +81,6 @@ def test_ekfac_type2(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -156,7 +155,6 @@ def test_ekfac_type2_weight_sharing(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -214,7 +212,6 @@ def test_ekfac_mc(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -230,7 +227,7 @@ def test_ekfac_mc(
 
     # Scale absolute tolerance by the number of outputs when using sum reduction.
     num_outputs = sum(y.numel() for _, y in data)
-    device_atol = 5e-3 if ekfac._infer_device() == device("cpu") else 1e-2
+    device_atol = 5e-3 if ekfac.device == device("cpu") else 1e-2
     atol = {"sum": device_atol * num_outputs, "mean": device_atol}[loss_func.reduction]
     rtol = {"sum": 2e-2, "mean": 2e-2}[loss_func.reduction]
 
@@ -290,7 +287,6 @@ def test_ekfac_mc_weight_sharing(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -308,7 +304,7 @@ def test_ekfac_mc_weight_sharing(
 
     # Scale absolute tolerance by the number of outputs when using sum reduction.
     num_outputs = sum(y.numel() for _, y in data)
-    device_atol = 5e-3 if ekfac._infer_device() == device("cpu") else 1e-2
+    device_atol = 5e-3 if ekfac.device == device("cpu") else 1e-2
     atol = {"sum": device_atol * num_outputs, "mean": device_atol}[loss_func.reduction]
     rtol = {"sum": 2e-2, "mean": 2e-2}[loss_func.reduction]
 
@@ -344,7 +340,6 @@ def test_ekfac_one_datum(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -389,7 +384,6 @@ def test_ekfac_mc_one_datum(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
     ekfac = EKFACLinearOperator(
         model,
@@ -438,7 +432,6 @@ def test_ekfac_ef_one_datum(
         data,
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
-        return_numpy=False,
     )
 
     ekfac = EKFACLinearOperator(