+++ changelog.md
@@ -8,8 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added/New
 
+### Fixed/Removed
+
+### Internal
+
+## [3.0.0] - 2025-10-16
+
+This new major release realizes all features described in our [position paper](https://arxiv.org/abs/2501.19183).
+Most importantly, **all linear operators are purely PyTorch by default now**.
+If you prefer working with SciPy linear operators (as was the default in `2.x`), you can simply call `.to_scipy()` on a linear operator.
+
+### Added/New
+
+- Add warning about overwriting model parameters when loading state dict in (E)KFAC ([PR](https://github.com/f-dangel/curvlinops/pull/196))
+
 - **Backward-incompatible:** Refactor class-based trace and diagonal estimators
-  into functions ([PR](https://github.com/f-dangel/curvlinops/pull/168)):
+  into functions ([PR](https://github.com/f-dangel/curvlinops/pull/168)) and assume
+  PyTorch instead of SciPy linear operators ([PR](https://github.com/f-dangel/curvlinops/pull/188)):
   - If you used `HutchinsonTraceEstimator`, switch to `hutchinson_trace`
   - If you used `HutchPPTraceEstimator`, switch to `hutchpp_trace`
   - If you used `HutchinsonDiagonalEstimator`, switch to `hutchinson_diag`
@@ -57,6 +72,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Switch CG implementation from SciPy to GPyTorch `CGInverseLinearOperator`
       ([PR](https://github.com/f-dangel/curvlinops/pull/190))
 
+- **Backward-incompatible** Refactor spectral density estimation methods to expect PyTorch linear operators and use PyTorch functions internally rather than SciPy/NumPy
+  ([PR](https://github.com/f-dangel/curvlinops/pull/194))
+
 - Add `EKFACLinearOperator` which implements EKFAC
   ([paper](https://arxiv.org/abs/1806.03884),
    [PR](https://github.com/f-dangel/curvlinops/pull/173))
@@ -80,6 +98,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (previously always on CPU)
   ([PR](https://github.com/f-dangel/curvlinops/pull/160))
 - Fixed the default value for `conlim` in `LSMRInverseLinearOperator` from `1e-8` to `1e8` ([PR](https://github.com/f-dangel/curvlinops/pull/180))
+- Bug in XDiag implementation that would only work with dense matrices, but not with
+  linear operators ([PR](https://github.com/f-dangel/curvlinops/pull/188))
 
 ### Internal
 
@@ -92,6 +112,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Test that two consecutive matrix-vector products of a linear operator match
   ([issue](https://github.com/f-dangel/curvlinops/issues/159),
    [PR](https://github.com/f-dangel/curvlinops/pull/175))
+- Reduce usage of SciPy in tests and examples
+  ([PR](https://github.com/f-dangel/curvlinops/pull/187))
+- Introduce `property`s for a linear operator's data type (`.dtype`) and device (`.device`)
+  ([PR](https://github.com/f-dangel/curvlinops/pull/189))
+- Move utility linear operators for examples into a single file
+  ([PR](https://github.com/f-dangel/curvlinops/pull/195))
 
 ## [2.0.1] - 2024-10-25
 
@@ -406,7 +432,8 @@ Adds various new features:
 
 Initial release
 
-[Unreleased]: https://github.com/f-dangel/curvlinops/compare/2.0.1...HEAD
+[Unreleased]: https://github.com/f-dangel/curvlinops/compare/3.0.0...HEAD
+[3.0.0]: https://github.com/f-dangel/curvlinops/releases/tag/3.0.0
 [2.0.1]: https://github.com/f-dangel/curvlinops/releases/tag/2.0.1
 [2.0.0]: https://github.com/f-dangel/curvlinops/releases/tag/2.0.0
 [1.2.0]: https://github.com/f-dangel/curvlinops/releases/tag/1.2.0